Extract local-function compiler fixtures
Move the complete local-function sample cluster into a focused file while preserving declarations, comments, and warning scopes.
